AI/requirements.txt
Package               Version
--------------------- --------
alembic               1.11.2
click                 8.1.6
colorama              0.4.6
Flask                 2.2.5
Flask-Migrate         4.0.4
flask-sqlalchemy      3.0.5
greenlet              2.0.2
importlib-metadata    6.7.0
importlib-resources   5.12.0
itsdangerous          2.1.2
Jinja2                3.1.2
joblib                1.3.1
Mako                  1.2.4
MarkupSafe            2.1.3
numpy                 1.21.6
opencv-contrib-python 3.4.2.16
opencv-python         3.4.2.16
Pillow                9.5.0
pip                   20.1.1
PyMySQL               1.1.0
scikit-learn          1.0.2
scipy                 1.7.3
setuptools            47.1.0
SQLAlchemy            2.0.19
threadpoolctl         3.1.0
typing-extensions     4.7.1
Werkzeug              2.2.3
zipp                  3.15.0
